add manual schematic placements
