+++ lib/manual-edits/index.ts
@@ -1,4 +1,5 @@
 export * from "./manual-edit-events"
 export * from "./manual_edit_file"
-export * from "./manual_pcb_position"
+export * from "./manual_pcb_placement"
+export * from "./manual_schematic_placement"
 export * from "./manual_trace_hint"
